requirements.txt
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
matplotlib
numpy
open3d==0.11.2
opencv-python==4.4.0.46
Pillow==8.1.0
scikit-image==0.17.2
scipy
